requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile
#
aiofiles==23.2.1
    # via gradio
altair==5.3.0
    # via gradio
annotated-types==0.6.0
    # via pydantic
anyio==4.3.0
    # via
    #   httpx
    #   starlette
    #   watchfiles
asttokens==2.4.1
    # via stack-data
attrs==23.2.0
    # via
    #   jsonschema
    #   referencing
certifi==2024.2.2
    # via
    #   httpcore
    #   httpx
    #   requests
charset-normalizer==3.3.2
    # via requests
click==8.1.7
    # via
    #   typer
    #   uvicorn
contourpy==1.2.1
    # via matplotlib
cycler==0.12.1
    # via matplotlib
decorator==5.1.1
    # via ipython
dnspython==2.6.1
    # via email-validator
email-validator==2.1.1
    # via fastapi
executing==2.0.1
    # via stack-data
fastapi==0.111.0
    # via
    #   fastapi-cli
    #   gradio
fastapi-cli==0.0.3
    # via fastapi
ffmpy==0.3.2
    # via gradio
filelock==3.14.0
    # via
    #   huggingface-hub
    #   torch
    #   transformers
fonttools==4.51.0
    # via matplotlib
fsspec==2024.3.1
    # via
    #   gradio-client
    #   huggingface-hub
    #   torch
gradio==4.14.0
    # via
    #   -r requirements.in
    #   gradio-imageslider
gradio-client==0.8.0
    # via gradio
gradio-imageslider==0.0.20
    # via -r requirements.in
h11==0.14.0
    # via
    #   httpcore
    #   uvicorn
httpcore==1.0.5
    # via httpx
httptools==0.6.1
    # via uvicorn
httpx==0.27.0
    # via
    #   fastapi
    #   gradio
    #   gradio-client
huggingface-hub==0.23.0
    # via
    #   -r requirements.in
    #   gradio
    #   gradio-client
    #   tokenizers
    #   transformers
idna==3.7
    # via
    #   anyio
    #   email-validator
    #   httpx
    #   requests
importlib-resources==6.4.0
    # via gradio
ipython==8.24.0
    # via -r requirements.in
jedi==0.19.1
    # via ipython
jinja2==3.1.4
    # via
    #   altair
    #   fastapi
    #   gradio
    #   torch
jsonschema==4.22.0
    # via altair
jsonschema-specifications==2023.12.1
    # via jsonschema
kiwisolver==1.4.5
    # via matplotlib
markdown-it-py==3.0.0
    # via rich
markupsafe==2.1.5
    # via
    #   gradio
    #   jinja2
matplotlib==3.8.4
    # via gradio
matplotlib-inline==0.1.7
    # via ipython
mdurl==0.1.2
    # via markdown-it-py
mpmath==1.3.0
    # via sympy
networkx==3.3
    # via torch
numpy==1.26.4
    # via
    #   altair
    #   contourpy
    #   gradio
    #   matplotlib
    #   opencv-python
    #   pandas
    #   torchvision
    #   transformers
opencv-python==4.9.0.80
    # via -r requirements.in
orjson==3.10.3
    # via
    #   fastapi
    #   gradio
packaging==24.0
    # via
    #   altair
    #   gradio
    #   gradio-client
    #   huggingface-hub
    #   matplotlib
    #   transformers
pandas==2.2.2
    # via
    #   altair
    #   gradio
parso==0.8.4
    # via jedi
pexpect==4.9.0
    # via ipython
pillow==10.3.0
    # via
    #   gradio
    #   gradio-imageslider
    #   matplotlib
    #   torchvision
prompt-toolkit==3.0.43
    # via ipython
ptyprocess==0.7.0
    # via pexpect
pure-eval==0.2.2
    # via stack-data
pydantic==2.7.1
    # via
    #   fastapi
    #   gradio
pydantic-core==2.18.2
    # via pydantic
pydub==0.25.1
    # via gradio
pygments==2.18.0
    # via
    #   ipython
    #   rich
pyparsing==3.1.2
    # via matplotlib
python-dateutil==2.9.0.post0
    # via
    #   matplotlib
    #   pandas
python-dotenv==1.0.1
    # via uvicorn
python-multipart==0.0.9
    # via
    #   fastapi
    #   gradio
pytz==2024.1
    # via pandas
pyyaml==6.0.1
    # via
    #   gradio
    #   huggingface-hub
    #   transformers
    #   uvicorn
referencing==0.35.1
    # via
    #   jsonschema
    #   jsonschema-specifications
regex==2024.5.10
    # via transformers
requests==2.31.0
    # via
    #   huggingface-hub
    #   transformers
rich==13.7.1
    # via typer
rpds-py==0.18.1
    # via
    #   jsonschema
    #   referencing
safetensors==0.4.3
    # via transformers
semantic-version==2.10.0
    # via gradio
shellingham==1.5.4
    # via typer
six==1.16.0
    # via
    #   asttokens
    #   python-dateutil
sniffio==1.3.1
    # via
    #   anyio
    #   httpx
stack-data==0.6.3
    # via ipython
starlette==0.37.2
    # via fastapi
sympy==1.12
    # via torch
tokenizers==0.19.1
    # via transformers
tomlkit==0.12.0
    # via gradio
toolz==0.12.1
    # via altair
torch==2.3.0
    # via
    #   -r requirements.in
    #   torchvision
torchvision==0.18.0
    # via -r requirements.in
tqdm==4.66.4
    # via
    #   huggingface-hub
    #   transformers
traitlets==5.14.3
    # via
    #   ipython
    #   matplotlib-inline
transformers==4.40.2
    # via -r requirements.in
typer[all]==0.12.3
    # via
    #   fastapi-cli
    #   gradio
typing-extensions==4.11.0
    # via
    #   fastapi
    #   gradio
    #   gradio-client
    #   huggingface-hub
    #   ipython
    #   pydantic
    #   pydantic-core
    #   torch
    #   typer
tzdata==2024.1
    # via pandas
ujson==5.10.0
    # via fastapi
urllib3==2.2.1
    # via requests
uvicorn[standard]==0.29.0
    # via
    #   fastapi
    #   fastapi-cli
    #   gradio
uvloop==0.19.0
    # via uvicorn
watchfiles==0.21.0
    # via uvicorn
wcwidth==0.2.13
    # via prompt-toolkit
websockets==11.0.3
    # via
    #   gradio-client
    #   uvicorn
